Fixed the SEO and document head
