+++ src/pages/_app.js
@@ -1,14 +1,12 @@
 import React from 'react';
 // import App from 'next/app'
 import { AnimatePresence } from 'framer-motion';
-import Head from 'components/head';
 import GlobalStyle from 'global.css.js';
 
 /* eslint-disable react/prop-types */
-const Main = ({ Component, pageProps, router }) => (
+const MyApp = ({ Component, pageProps, router }) => (
   <>
     <GlobalStyle />
-    <Head />
     <AnimatePresence exitBeforeEnter>
       <Component {...pageProps} key={router.route} />
     </AnimatePresence>
@@ -20,11 +18,11 @@ const Main = ({ Component, pageProps, router }) => (
 // perform automatic static optimization, causing every page in your app to
 // be server-side rendered.
 //
-// Main.getInitialProps = async (appContext) => {
+// MyApp.getInitialProps = async (appContext) => {
 //   // calls page's `getInitialProps` and fills `appProps.pageProps`
 //   const appProps = await App.getInitialProps(appContext);
 //
 //   return { ...appProps }
 // }
 
-export default Main;
+export default MyApp;
